pyproject.toml
[project]
name = "openapi-mcp-server"
version = "0.1.0"
description = "MCP server for any OpenAPI spec."
authors = [
  { name = "Aliaksei Yatsau", email = "yatsew.alexey@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-dotenv>=1.0.1,<2.0.0",
  "mcp[cli]",
  "requests>=2.31.0,<3.0.0",
  "prance>=0.22.0,<0.23.0",
  "PyYAML>=6.0,<7.0",
  "openapi-spec-validator>=0.7.1,<0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "ruff>=0.4.0,<0.5.0"
]

[tool.uv]
# uv-specific configuration (if any)

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = [
    "src"
]
testpaths = [
    "tests"
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 140
src = ["src", "tests"]
exclude = ["__init__.py"]

[tool.ruff.lint]
select = ["E", "F", "I"]
# E: pycodestyle errors, F: pyflakes, I: isort